Moving the transient Weather class out of model and into the DAO where it's actually more relevant
